requirements.txt
streamlit==1.13.0
pandas==1.3.5
scikit-learn==1.0.2
yfinance==0.1.74
requests~=2.28.1
numpy~=1.21.6
matplotlib~=3.5.3
plotly~=5.10.0
pytickersymbols~=1.8.78
python-dateutil~=2.8.2